refactor : add url launcher inside helper functions folder
