+++ lib/core/api_manger/api_service.dart
@@ -2,8 +2,8 @@ import 'package:dio/dio.dart';
 import 'package:retrofit/error_logger.dart';
 import 'package:retrofit/http.dart';
 import 'package:tracking_app/core/utils/constants.dart';
-import 'package:tracking_app/features/apply/data/models/apply_model/apply_response/apply_response.dart';
-import 'package:tracking_app/features/apply/data/models/get_all_vehicles_response/get_all_vehicles_response.dart';
+import 'package:tracking_app/features/auth/apply/data/models/apply_model/apply_response/apply_response.dart';
+import 'package:tracking_app/features/auth/apply/data/models/get_all_vehicles_response/get_all_vehicles_response.dart';
 import 'package:tracking_app/features/auth/login/data/model/login_request.dart';
 import 'package:tracking_app/features/auth/login/data/model/login_response.dart';
 import 'package:tracking_app/features/auth/forget_password/data/models/request/forget_password_request.dart';
