+++ lib/features/auth/apply/data/models/get_all_vehicles_response/vehicle.dart
