+++ lib/features/auth/apply/domain/use_case/apply_use_case.dart
@@ -1,8 +1,8 @@
 import 'package:injectable/injectable.dart';
 import 'package:tracking_app/core/base/api_result.dart';
-import 'package:tracking_app/features/apply/data/models/apply_model/apply_data.dart';
-import 'package:tracking_app/features/apply/data/models/apply_model/apply_response/apply_response.dart';
-import 'package:tracking_app/features/apply/domain/repos/auth_repo.dart';
+import 'package:tracking_app/features/auth/apply/data/models/apply_model/apply_data.dart';
+import 'package:tracking_app/features/auth/apply/data/models/apply_model/apply_response/apply_response.dart';
+import 'package:tracking_app/features/auth/apply/domain/repos/auth_repo.dart';
 
 @injectable
 class ApplyUseCase {
