+++ lib/features/auth/apply/domain/use_case/get_vehicle_use_case.dart
@@ -1,7 +1,7 @@
 import 'package:injectable/injectable.dart';
 import 'package:tracking_app/core/base/api_result.dart';
-import 'package:tracking_app/features/apply/data/models/get_all_vehicles_response/get_all_vehicles_response.dart';
-import 'package:tracking_app/features/apply/domain/repos/auth_repo.dart';
+import 'package:tracking_app/features/auth/apply/data/models/get_all_vehicles_response/get_all_vehicles_response.dart';
+import 'package:tracking_app/features/auth/apply/domain/repos/auth_repo.dart';
 
 @injectable
 class GetVehiclesUseCase {
