+++ lib/features/auth/apply/presentation/view/apply_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:tracking_app/core/di/di.dart';
-import 'package:tracking_app/features/apply/domain/use_case/apply_use_case.dart';
-import 'package:tracking_app/features/apply/domain/use_case/get_vehicle_use_case.dart';
-import 'package:tracking_app/features/apply/presentation/view/widgets/apply_screen_body.dart';
-import 'package:tracking_app/features/apply/presentation/view_model/cubit/apply_cubit.dart';
+import 'package:tracking_app/features/auth/apply/domain/use_case/apply_use_case.dart';
+import 'package:tracking_app/features/auth/apply/domain/use_case/get_vehicle_use_case.dart';
+import 'package:tracking_app/features/auth/apply/presentation/view/widgets/apply_screen_body.dart';
+import 'package:tracking_app/features/auth/apply/presentation/view_model/cubit/apply_cubit.dart';
 
 
 class ApplyScreen extends StatelessWidget {
