+++ lib/features/auth/apply/presentation/view/widgets/welcome_text_in_apply_view.dart
