+++ lib/main.dart
@@ -1,6 +1,6 @@
 // import 'package:bloc/bloc.dart';
 import 'package:easy_localization/easy_localization.dart';
-import 'package:firebase_core/firebase_core.dart';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:provider/provider.dart';
@@ -16,15 +16,16 @@ import 'package:tracking_app/features/auth/login/presentation/view_model/login_c
 import 'package:tracking_app/features/auth/login/domain/usecases/login_usecase.dart';
 import 'package:tracking_app/features/auth/login/presentation/view_model/login_cubit.dart';
 import 'package:tracking_app/generated/codegen_loader.g.dart';
-
+import 'package:firebase_core/firebase_core.dart';
 import 'firebase_options.dart';
+
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await EasyLocalization.ensureInitialized();
   configureDependencies();
   Bloc.observer = SimpleBlocObserver();
-  await Firebase.initializeApp(
+  await  Firebase.initializeApp(
     options: DefaultFirebaseOptions.currentPlatform,
   );
