+++ pubspec.yaml
@@ -52,6 +52,7 @@ dependencies:
     sdk: flutter
   step_progress_indicator: ^1.0.2
   firebase_core: ^3.13.0
+  
 
 dev_dependencies:
   build_runner: ^2.3.3
